refactor(upskill): extract recommendations and tabs discovery into separ
Agent-Id: agent-66d666fb-8feb-4d18-9419-1c6110f275e5
Linked-Note-Id: 64e743ef-5824-4166-9c64-385efa16b221
